Phase 2: Move core protocol and engine abstraction
- Move deployments/protocol.py → core/protocol.py
- Move deployments/llm/llm_engine.py → core/engine/protocol.py
- Update all imports across codebase (6 files updated)
